requirements.txt
flask==2.3.0
Werkzeug==2.3.8
gunicorn==21.2.0
setuptools==75.8.0
wheel==0.45.0
python-dotenv==1.2.1
bcrypt==5.0.0
pyodbc==5.3.0
pyjwt==2.11.0
